feat: /comments  - 댓글 구현

+++ src/main/java/com/efub/lakkulakku/domain/comment/controller/CommentController.java
@@ -1,13 +1,18 @@
 package com.efub.lakkulakku.domain.comment.controller;
 
+import com.efub.lakkulakku.domain.comment.dto.CommentDeleteReqDto;
+import com.efub.lakkulakku.domain.comment.dto.CommentReqDto;
 import com.efub.lakkulakku.domain.comment.dto.CommentResDto;
+import com.efub.lakkulakku.domain.comment.dto.CommentUpdateResDto;
 import com.efub.lakkulakku.domain.comment.exception.CommentNotFoundException;
 import com.efub.lakkulakku.domain.comment.exception.ParentNotFoundException;
 import com.efub.lakkulakku.domain.comment.entity.Comment;
+import com.efub.lakkulakku.domain.comment.exception.UnauthorizedException;
 import com.efub.lakkulakku.domain.comment.repository.CommentRepository;
 import com.efub.lakkulakku.domain.comment.service.CommentService;
 import com.efub.lakkulakku.domain.diary.exception.DiaryNotFoundException;
 import com.efub.lakkulakku.domain.diary.repository.DiaryRepository;
+import com.efub.lakkulakku.domain.likes.dto.LikeResDto;
 import com.efub.lakkulakku.domain.users.entity.Users;
 import com.efub.lakkulakku.domain.users.service.AuthUsers;
 import lombok.RequiredArgsConstructor;
@@ -19,6 +24,8 @@ import org.springframework.web.bind.annotation.*;
 import java.time.LocalDate;
 import java.util.UUID;
 
+import static com.efub.lakkulakku.global.constant.ResponseConstant.COMMENT_DELETE_SUCCESS;
+
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/diaries")
@@ -29,53 +36,37 @@ public class CommentController {
 	private final DiaryRepository diaryRepository;
 
 	@PostMapping("/{date}/comments")
-	@ResponseStatus(HttpStatus.CREATED)
-	public ResponseEntity<?> commentAdd(@AuthUsers Users user, @PathVariable("date") @DateTimeFormat(pattern = "yyyy-MM-dd") LocalDate date, /*UUID id, UUID parentId,*/ @RequestBody CommentResDto commentResDto) {
+	public ResponseEntity<?> commentAdd(@AuthUsers Users user, @PathVariable("date") @DateTimeFormat(pattern = "yyyy-MM-dd") LocalDate date,@RequestBody CommentReqDto commentReqDto) {
 
 		if (!diaryRepository.existsByDate(date))
 			throw new DiaryNotFoundException();
+		if (commentReqDto.getParentId() != null) {
+			commentRepository.findById(commentReqDto.getParentId())
+					.orElseThrow(() -> new ParentNotFoundException());
+		}
 
-		commentService.addComment(commentResDto);
+		CommentResDto commentResDto = commentService.addComment(user, date, commentReqDto);
 
-		return ResponseEntity.ok("해당 댓글이 작성되었습니다.");
+		return ResponseEntity.ok(commentResDto);
 	}
 
 	@DeleteMapping("/{date}/comments/{id}")
-	public ResponseEntity<?> commentRemove(@AuthUsers Users user, @PathVariable("date") @DateTimeFormat(pattern = "yyyy-MM-dd") LocalDate date,@PathVariable("id") UUID id, @RequestBody CommentResDto commentResDto) {
+	public ResponseEntity<?> commentRemove(@AuthUsers Users user, @PathVariable("date") @DateTimeFormat(pattern = "yyyy-MM-dd") LocalDate date, @PathVariable("id") UUID id) {
 
-		Comment comment = commentRepository.findById(id)
-				.orElseThrow(CommentNotFoundException::new);
+		if (!commentRepository.findById(id).get().getUsers().equals(user))
+			throw new UnauthorizedException();
 
 		commentService.removeComment(id);
 
-		return ResponseEntity.ok("댓글이 삭제되었습니다.");
+		return ResponseEntity.ok(COMMENT_DELETE_SUCCESS);
 	}
 
 	@PutMapping("/{date}/comments/{id}")
-	public ResponseEntity update(@AuthUsers Users user, @PathVariable("date") @DateTimeFormat(pattern = "yyyy-MM-dd") LocalDate date,@PathVariable("id") UUID id, @RequestBody CommentResDto commentResDto) {
+	public ResponseEntity<?> update(@AuthUsers Users user, @PathVariable("date") @DateTimeFormat(pattern = "yyyy-MM-dd") LocalDate date,@PathVariable("id") UUID id, @RequestBody CommentReqDto commentReqDto) {
 
-		commentService.update(id, commentResDto);
+		CommentUpdateResDto commentUpdateResDto = commentService.update(user, id, date, commentReqDto);
 
-		return ResponseEntity.ok(CommentUpdateResDto);
-	}
-
-	@PostMapping("/{date}/comments/{id}")
-	public ResponseEntity<?> recommentAdd(@AuthUsers Users user, @PathVariable("date") @DateTimeFormat(pattern = "yyyy-MM-dd") LocalDate date,@PathVariable("id") UUID id, UUID parentId, @RequestBody CommentResDto commentResDto) {
-
-		//Comment comment = commentRepository.findById(id)
-		//		.orElseThrow(ParentNotFoundException::new);
-
-		if (!diaryRepository.existsByDate(date))
-			throw new DiaryNotFoundException();
-
-		Comment comment = commentRepository.findByParentId(parentId)
-		      .orElseThrow(ParentNotFoundException::new);
-
-		//commentService.addRecomment(commentResDto.getId(), commentResDto.getParentId(), commentResDto);
-
-		commentService.addRecomment(id, parentId, commentResDto);
-
-		return ResponseEntity.ok("해당 댓글이 작성되었습니다.");
+		return ResponseEntity.ok(commentUpdateResDto);
 	}
 
 }
